src/main/webapp/app/entities/candidature-etudiant/list/candidature-etudiant.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpHeaders, HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';

import { CandidatureEtudiantService } from '../service/candidature-etudiant.service';

import { CandidatureEtudiantComponent } from './candidature-etudiant.component';

describe('CandidatureEtudiant Management Component', () => {
  let comp: CandidatureEtudiantComponent;
  let fixture: ComponentFixture<CandidatureEtudiantComponent>;
  let service: CandidatureEtudiantService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [CandidatureEtudiantComponent],
    })
      .overrideTemplate(CandidatureEtudiantComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(CandidatureEtudiantComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(CandidatureEtudiantService);

    const headers = new HttpHeaders();
    jest.spyOn(service, 'query').mockReturnValue(
      of(
        new HttpResponse({
          body: [{ id: 123 }],
          headers,
        })
      )
    );
  });

  it('Should call load all on init', () => {
    // WHEN
    comp.ngOnInit();

    // THEN
    expect(service.query).toHaveBeenCalled();
    expect(comp.candidatureEtudiants[0]).toEqual(expect.objectContaining({ id: 123 }));
  });

  it('should load a page', () => {
    // WHEN
    comp.loadPage(1);

    // THEN
    expect(service.query).toHaveBeenCalled();
    expect(comp.candidatureEtudiants[0]).toEqual(expect.objectContaining({ id: 123 }));
  });

  it('should calculate the sort attribute for an id', () => {
    // WHEN
    comp.ngOnInit();

    // THEN
    expect(service.query).toHaveBeenCalledWith(expect.objectContaining({ sort: ['id,asc'] }));
  });

  it('should calculate the sort attribute for a non-id attribute', () => {
    // INIT
    comp.ngOnInit();

    // GIVEN
    comp.predicate = 'name';

    // WHEN
    comp.loadPage(1);

    // THEN
    expect(service.query).toHaveBeenLastCalledWith(expect.objectContaining({ sort: ['name,asc', 'id'] }));
  });

  it('should re-initialize the page', () => {
    // WHEN
    comp.loadPage(1);
    comp.reset();

    // THEN
    expect(comp.page).toEqual(0);
    expect(service.query).toHaveBeenCalledTimes(2);
    expect(comp.candidatureEtudiants[0]).toEqual(expect.objectContaining({ id: 123 }));
  });
});
